+++ build-tools/generator/generator.js
@@ -97,11 +97,11 @@ class PapertrainGenerator
 
                 spinner.start();
                 spinner.text = 'Creating directory';
-                await this.generateDirectory(path);
+                await this.generateDirectory(path, type);
             }
             else
             {
-                path = `${ basePath }/_globals`;
+                path = `${ basePath }/_global-stylesheets`;
                 spinner.start();
             }
 
@@ -174,10 +174,10 @@ class PapertrainGenerator
             .replace(new RegExp(/\w/), s => s.toUpperCase());
     }
 
-    generateDirectory(path)
+    generateDirectory(path, type)
     {
         return new Promise((resolve, reject)=>{
-            fs.mkdir(`${ path }/${ this.kebabName }`, (err)=>{
+            fs.mkdir(`${ path }/${ (type === 'Web Component') ? '_' : '' }${ this.kebabName }`, (err)=>{
                 if (err)
                 {
                     spinner.fail();
@@ -211,9 +211,9 @@ class PapertrainGenerator
                 {
                     try
                     {
-                        await this.generateComponent(path);
-                        await this.generateStylesheet(path);
-                        await this.generateScript(path);
+                        await this.generateComponent(path, type);
+                        await this.generateStylesheet(path, type);
+                        await this.generateScript(path, type);
                         resolve();
                     }
                     catch (error)
@@ -280,16 +280,16 @@ class PapertrainGenerator
         });
     }
 
-    generateStylesheet(path)
+    generateStylesheet(path, type)
     {
         return new Promise((resolve, reject)=>{
-            fs.copyFile(`${ __dirname }/files/stylesheet`, `${ path }/${ this.kebabName }/${ this.kebabName }.${ StylesheetFileType }`, (error)=>{
+            fs.copyFile(`${ __dirname }/files/stylesheet`, `${ path }/${ (type === 'Web Component') ? '_' : '' }${ this.kebabName }/${ this.kebabName }.${ StylesheetFileType }`, (error)=>{
                 if (error)
                 {
                     reject(error);
                 }
 
-                fs.readFile(`${ path }/${ this.kebabName }/${ this.kebabName }.${ StylesheetFileType }`, 'utf-8', (error, data)=>{
+                fs.readFile(`${ path }/${ (type === 'Web Component') ? '_' : '' }${ this.kebabName }/${ this.kebabName }.${ StylesheetFileType }`, 'utf-8', (error, data)=>{
                     if (error)
                     {
                         reject(error);
@@ -298,7 +298,7 @@ class PapertrainGenerator
                     let modifiedFile = data;
                     modifiedFile = modifiedFile.replace(/REPLACE_WITH_CSS/g, this.cssName);
 
-                    fs.writeFile(`${ path }/${ this.kebabName }/${ this.kebabName }.${ StylesheetFileType }`, modifiedFile, 'utf-8', (err)=>{
+                    fs.writeFile(`${ path }/${ (type === 'Web Component') ? '_' : '' }${ this.kebabName }/${ this.kebabName }.${ StylesheetFileType }`, modifiedFile, 'utf-8', (err)=>{
                         if (err)
                         {
                             spinner.fail();
@@ -315,13 +315,13 @@ class PapertrainGenerator
     generateComponent(path)
     {
         return new Promise((resolve, reject)=>{
-            fs.copyFile(`${ __dirname }/files/component`, `${ path }/${ this.kebabName }/index.${ HTMLFileType }`, (error)=>{
+            fs.copyFile(`${ __dirname }/files/component`, `${ path }/_${ this.kebabName }/index.${ HTMLFileType }`, (error)=>{
                 if (error)
                 {
                     reject(error);
                 }
 
-                fs.readFile(`${ path }/${ this.kebabName }/index.${ HTMLFileType }`, 'utf-8', (error, data)=>{
+                fs.readFile(`${ path }/_${ this.kebabName }/index.${ HTMLFileType }`, 'utf-8', (error, data)=>{
                     if (error)
                     {
                         reject(error);
@@ -330,7 +330,7 @@ class PapertrainGenerator
                     let modifiedFile = data;
                     modifiedFile = modifiedFile.replace(/REPLACE_WITH_KEBAB/g, this.kebabName);
 
-                    fs.writeFile(`${ path }/${ this.kebabName }/index.${ HTMLFileType }`, modifiedFile, 'utf-8', (err)=>{
+                    fs.writeFile(`${ path }/_${ this.kebabName }/index.${ HTMLFileType }`, modifiedFile, 'utf-8', (err)=>{
                         if (err)
                         {
                             spinner.fail();
@@ -347,13 +347,13 @@ class PapertrainGenerator
     generateScript(path)
     {
         return new Promise((resolve, reject)=>{
-            fs.copyFile(`${ __dirname }/files/script`, `${ path }/${ this.kebabName }/${ this.kebabName }.${ this.scriptFileType }`, (error)=>{
+            fs.copyFile(`${ __dirname }/files/script`, `${ path }/_${ this.kebabName }/${ this.kebabName }.${ this.scriptFileType }`, (error)=>{
                 if (error)
                 {
                     reject(error);
                 }
 
-                fs.readFile(`${ path }/${ this.kebabName }/${ this.kebabName }.${ this.scriptFileType }`, 'utf-8', (error, data)=>{
+                fs.readFile(`${ path }/_${ this.kebabName }/${ this.kebabName }.${ this.scriptFileType }`, 'utf-8', (error, data)=>{
                     if (error)
                     {
                         reject(error);
@@ -365,7 +365,7 @@ class PapertrainGenerator
                     const devName = (process.env.DEV_NAME) ? process.env.DEV_NAME : 'Anonymous';
                     modifiedFile = modifiedFile.replace(/AUTHOR_NAME/g, devName);
 
-                    fs.writeFile(`${ path }/${ this.kebabName }/${ this.kebabName }.${ this.scriptFileType }`, modifiedFile, 'utf-8', (err)=>{
+                    fs.writeFile(`${ path }/_${ this.kebabName }/${ this.kebabName }.${ this.scriptFileType }`, modifiedFile, 'utf-8', (err)=>{
                         if (err)
                         {
                             spinner.fail();